refactor the heap and binary heap
in order to have trait object safety: https://doc.rust-lang.org/reference/items/traits.html#object-safety
Otherwise we can not reference this trait in another struct.
Also refactor the `heap` member to be `data`, it seems to be less
confusing as we are not really containing a heap, but a
datastructure allowing us to store the heap (anyway, this is really
minor).

Because of that we don't have a fluent API call when doing `add`.
Need to check if there are other solutions that would allow this.