implements a KHeap
A KHeap is a heap storing the k-maximum values, or k-minimum values.
The values are unsorted, so extracting the max or the min is O(k),
but extracting the k-min or k-max is just checking the head of the
heap, so O(1).
For storing the k-maximum values we use a min-heap, and a max-heap
for the k-minimum values. Like that the k-th maximum value or k-th
minimum value is always on the top of the heap, and we easily now if
the new value to push should be part of those k values.

wip:
- as an exercise, we could implement iterator in order to be able to
  ensure fully that the KHeap is containing the expected values.