implements iterator for binary heap
But struggling to implement the same mechanism for the `KHeap`.
The goal would have been to be able to write a test like this:
```rust

    #[test]
    fn it_should_contain_3_min_after_n_pushes() {
        // GIVEN
        let mut k_heap = KHeap::min_with_size(3);
        k_heap.push(3);
        k_heap.push(0);
        k_heap.push(1);
        k_heap.push(18);
        k_heap.push(-11);
        k_heap.push(-1);
        k_heap.push(42);

        // WHEN
        let mut iterated = vec![];
        for e in &k_heap {
            iterated.push(e);
        }

        // THEN
        assert_eq!(iterated, vec![-11, -1, 0]);
    }
```
The issue is that we can not delegate to the wrapped heap as the
`Heap` trait doesn't implement `Iterator`.
So I guess the solution would be to make the `Heap` trait be
implementing `Iterator`.

Currently this is not exactly `BinaryHeap` which is implementing
the iterator trait, but `&BinaryHeap`, because if I make the type
`BinaryHeap` being implementing the iterator then I can not use the
heap anymore after the iteration because it has been move...
I definitely have still a lot to learn about those concept of
borrowing/moving/ownership... I'm too often trying to guess code :/