adds solution for logger rate limiter
This is a naive solution, the map would grow as it would retain one
entry per type of message.
In a real life example, we would have multiple solution, debouncing
the log, so that the log is printed only after 10s, and then while
printing the log we could also remove the entry.
We could also have some "smart" cache with automatic entry removal
if the timestamp is older than 10s.
...