adds solution to challenge guess the word
This is a basic solution, the problem is that it might fail, or
not...
This is not predictable, the commented test work about 9 times out
of 10 tries, it will just depend of the order of the candidates we
are looking at.

So one solution to optimize this could be to find a way to pick a
candidate that will discriminate more candidates.