fix(challenge): fix guess the word rust code
The pop method was not used, because I was messing with borrowing
and ownership and didn't manage to make it work. So I just gave it a
second try, and this time this is working.
I would still need to improve my knowledge, because sometimes I have
the feeling that I'm guessing instead of knowing what to do, this is
a bit frustrating, but I guess this is part of the process...

There is still one thing I don't fully understand, why do I need the
cloned inside the `pop` method. The compiler is complaining like
that:
```
error[E0502]: cannot borrow `*set` as mutable because it is also borrowed as immutable
  --> src/challenge/guess_the_word.rs:43:5
   |
42 |     let elem = set.iter().next().unwrap();
   |                --- immutable borrow occurs here
43 |     set.take(elem).unwrap()
   |     ^^^^----^^^^^^
   |     |   |
   |     |   immutable borrow later used by call
   |     mutable borrow occurs here
```
But I was under the impression that the immutable borrowing was done
after the `iter().next()` would be executed, so that we could do a
mutable borrowing later on... But it seems we can't.