implements a really naive solution
This is using a sort of DFS to discover the islands and to erase
them from the map. I have the feeling that this is doing a bit of
unnecessary check, because we are revisiting often nodes.
For example if we have only one island taking the full map, we will
visit the map once to discover the island and erase it and then
revisit the whole map to see if there isn't any island.

There is probably something to do with an union find, but too late
to make my brain work and implement that... Will see later.